Removed preloader and added calendly api to set an appointment
Since there's a lot of pictures on the website on the front page
it sometimes take a lot of time to preload first before rendering the
page to the user. After 3 seconds of loading a user usually thinks
there's a problem with the webpage or the connection or he/she's
just an impatient bastard and leave the page. To prevent that, I removed
the preloader and the user will see the page showing as it's downloading the data.

Also added calendly to set appointment.

Will do
 - Prepopulate the field of calendly
 - Add square and paypal
 - work on the user interface login
 - Fix minor error and the design